platform/bootstrap/antsctl.py
"""
ANTS Control CLI (antsctl).
Bootstrap AI Agent for self-deploying ANTS infrastructure.
"""
import click
import yaml
import asyncio
from pathlib import Path
from typing import Dict, Any, Optional
import structlog

logger = structlog.get_logger()


@click.group()
@click.option('--verbose', '-v', is_flag=True, help='Enable verbose output')
@click.option('--config', '-c', type=click.Path(), help='Path to ants-spec.yaml')
@click.pass_context
def cli(ctx: click.Context, verbose: bool, config: Optional[str]):
    """ANTS Control CLI - AI-Agent Native Tactical System."""
    ctx.ensure_object(dict)
    ctx.obj['verbose'] = verbose
    ctx.obj['config_path'] = config


@cli.command()
@click.argument('spec_file', type=click.Path(exists=True))
@click.option('--dry-run', is_flag=True, help='Show what would be deployed')
@click.option('--environment', '-e', default='dev', help='Target environment')
@click.pass_context
def deploy(ctx: click.Context, spec_file: str, dry_run: bool, environment: str):
    """Deploy ANTS from specification file."""
    click.echo(f"Deploying ANTS from {spec_file} to {environment}...")

    spec = load_spec(spec_file)

    if ctx.obj['verbose']:
        click.echo(f"Loaded spec: {spec}")

    if dry_run:
        click.echo("\n=== DRY RUN - No changes will be made ===\n")
        show_deployment_plan(spec, environment)
    else:
        asyncio.run(execute_deployment(spec, environment))


@cli.command()
@click.argument('spec_file', type=click.Path(exists=True))
@click.pass_context
def validate(ctx: click.Context, spec_file: str):
    """Validate ANTS specification file."""
    click.echo(f"Validating {spec_file}...")

    try:
        spec = load_spec(spec_file)
        errors = validate_spec(spec)

        if errors:
            click.echo(click.style("Validation FAILED:", fg='red'))
            for error in errors:
                click.echo(f"  - {error}")
            raise click.Abort()
        else:
            click.echo(click.style("Validation PASSED", fg='green'))

    except Exception as e:
        click.echo(click.style(f"Error: {e}", fg='red'))
        raise click.Abort()


@cli.command()
@click.option('--environment', '-e', default='dev', help='Target environment')
@click.pass_context
def status(ctx: click.Context, environment: str):
    """Show ANTS deployment status."""
    click.echo(f"ANTS Status for {environment}:\n")

    status_data = asyncio.run(get_deployment_status(environment))

    # Display status table
    click.echo(f"{'Component':<25} {'Status':<15} {'Version':<10} {'Replicas':<10}")
    click.echo("-" * 60)

    for component, info in status_data.items():
        status_color = 'green' if info['status'] == 'healthy' else 'red'
        status_str = click.style(info['status'], fg=status_color)
        click.echo(f"{component:<25} {status_str:<24} {info.get('version', 'N/A'):<10} {info.get('replicas', 'N/A'):<10}")


@cli.command()
@click.argument('agent_type')
@click.option('--replicas', '-r', type=int, help='Number of replicas')
@click.option('--environment', '-e', default='dev', help='Target environment')
@click.pass_context
def scale(ctx: click.Context, agent_type: str, replicas: int, environment: str):
    """Scale agent workers."""
    click.echo(f"Scaling {agent_type} to {replicas} replicas in {environment}...")
    asyncio.run(scale_agents(agent_type, replicas, environment))
    click.echo(click.style("Scaling complete", fg='green'))


@cli.command()
@click.argument('component')
@click.option('--follow', '-f', is_flag=True, help='Follow log output')
@click.option('--lines', '-n', type=int, default=100, help='Number of lines')
@click.pass_context
def logs(ctx: click.Context, component: str, follow: bool, lines: int):
    """View component logs."""
    click.echo(f"Logs for {component}:")

    if follow:
        asyncio.run(stream_logs(component))
    else:
        log_data = asyncio.run(get_logs(component, lines))
        for line in log_data:
            click.echo(line)


@cli.group()
def agent():
    """Agent management commands."""
    pass


@agent.command('list')
@click.option('--environment', '-e', default='dev', help='Target environment')
def agent_list(environment: str):
    """List registered agents."""
    click.echo(f"Agents in {environment}:\n")

    agents = asyncio.run(get_agents(environment))

    click.echo(f"{'Type':<30} {'Name':<25} {'Status':<15}")
    click.echo("-" * 70)

    for agent in agents:
        status_color = 'green' if agent['status'] == 'ready' else 'yellow'
        status_str = click.style(agent['status'], fg=status_color)
        click.echo(f"{agent['type']:<30} {agent['name']:<25} {status_str}")


@agent.command('invoke')
@click.argument('agent_type')
@click.option('--input', '-i', 'input_file', type=click.Path(exists=True), help='Input JSON file')
@click.option('--tenant', '-t', required=True, help='Tenant ID')
def agent_invoke(agent_type: str, input_file: Optional[str], tenant: str):
    """Invoke an agent."""
    import json

    input_data = {}
    if input_file:
        with open(input_file) as f:
            input_data = json.load(f)

    click.echo(f"Invoking {agent_type} for tenant {tenant}...")

    result = asyncio.run(invoke_agent(agent_type, input_data, tenant))

    click.echo("\nResult:")
    click.echo(json.dumps(result, indent=2))


@cli.group()
def memory():
    """Memory substrate commands."""
    pass


@memory.command('search')
@click.argument('query')
@click.option('--type', '-t', 'memory_type', default='semantic', help='Memory type')
@click.option('--tenant', required=True, help='Tenant ID')
@click.option('--limit', '-n', type=int, default=10, help='Result limit')
def memory_search(query: str, memory_type: str, tenant: str, limit: int):
    """Search agent memory."""
    click.echo(f"Searching {memory_type} memory for: {query}\n")

    results = asyncio.run(search_memory(query, memory_type, tenant, limit))

    for i, result in enumerate(results, 1):
        click.echo(f"{i}. Score: {result.get('score', 0):.3f}")
        click.echo(f"   {result.get('content', '')[:100]}...")
        click.echo()


@cli.group()
def metrics():
    """Metrics and monitoring commands."""
    pass


@metrics.command('clear')
@click.option('--environment', '-e', default='dev', help='Target environment')
def metrics_clear(environment: str):
    """Show CLEAR metrics (Cost, Latency, Efficacy, Assurance, Reliability)."""
    click.echo(f"CLEAR Metrics for {environment}:\n")

    data = asyncio.run(get_clear_metrics(environment))

    click.echo("Cost:")
    click.echo(f"  Total Tokens: {data['cost']['total_tokens']:,}")
    click.echo(f"  Estimated USD: ${data['cost']['estimated_cost_usd']:.2f}")

    click.echo("\nLatency:")
    click.echo(f"  P50: {data['latency']['p50_ms']}ms")
    click.echo(f"  P95: {data['latency']['p95_ms']}ms")
    click.echo(f"  P99: {data['latency']['p99_ms']}ms")

    click.echo("\nEfficacy:")
    click.echo(f"  Success Rate: {data['efficacy']['success_rate']:.1%}")
    click.echo(f"  Avg Confidence: {data['efficacy']['avg_confidence']:.1%}")

    click.echo("\nAssurance:")
    click.echo(f"  Policy Compliance: {data['assurance']['policy_compliance_rate']:.1%}")
    click.echo(f"  Audit Coverage: {data['assurance']['audit_coverage']:.1%}")

    click.echo("\nReliability:")
    click.echo(f"  Uptime: {data['reliability']['uptime']:.3%}")
    click.echo(f"  Error Rate: {data['reliability']['error_rate']:.1%}")


# Helper functions
def load_spec(spec_file: str) -> Dict[str, Any]:
    """Load ANTS specification from YAML file."""
    with open(spec_file) as f:
        return yaml.safe_load(f)


def validate_spec(spec: Dict[str, Any]) -> list:
    """Validate ANTS specification."""
    errors = []

    # Required fields
    required = ['version', 'tenant', 'regions', 'modules']
    for field in required:
        if field not in spec:
            errors.append(f"Missing required field: {field}")

    # Validate modules
    if 'modules' in spec:
        valid_modules = ['anf', 'aks', 'databricks', 'ai_foundry', 'agents']
        for module in spec['modules']:
            if module not in valid_modules:
                errors.append(f"Unknown module: {module}")

    return errors


def show_deployment_plan(spec: Dict[str, Any], environment: str):
    """Show what would be deployed."""
    click.echo("Deployment Plan:")
    click.echo(f"  Tenant: {spec.get('tenant', {}).get('name', 'Unknown')}")
    click.echo(f"  Environment: {environment}")
    click.echo(f"  Regions: {spec.get('regions', [])}")
    click.echo(f"  Modules: {spec.get('modules', [])}")


async def execute_deployment(spec: Dict[str, Any], environment: str):
    """Execute deployment based on spec."""
    logger.info("executing_deployment", environment=environment)
    # Implementation would provision infrastructure
    click.echo("Deployment initiated...")


async def get_deployment_status(environment: str) -> Dict[str, Any]:
    """Get deployment status."""
    return {
        "api-gateway": {"status": "healthy", "version": "1.0.0", "replicas": 3},
        "orchestrator": {"status": "healthy", "version": "1.0.0", "replicas": 2},
        "agent-workers": {"status": "healthy", "version": "1.0.0", "replicas": 5},
        "policy-engine": {"status": "healthy", "version": "0.60.0", "replicas": 2},
        "memory-substrate": {"status": "healthy", "version": "1.0.0", "replicas": 2}
    }


async def scale_agents(agent_type: str, replicas: int, environment: str):
    """Scale agent workers."""
    logger.info("scaling_agents", agent_type=agent_type, replicas=replicas)


async def stream_logs(component: str):
    """Stream logs for component."""
    pass


async def get_logs(component: str, lines: int) -> list:
    """Get logs for component."""
    return [f"Log line {i}" for i in range(lines)]


async def get_agents(environment: str) -> list:
    """Get registered agents."""
    return [
        {"type": "finance.reconciliation", "name": "Reconciliation Agent", "status": "ready"},
        {"type": "retail.inventory", "name": "Inventory Agent", "status": "ready"},
        {"type": "cybersecurity.defender", "name": "Defender Triage Agent", "status": "ready"},
        {"type": "selfops.infra", "name": "InfraOps Agent", "status": "ready"}
    ]


async def invoke_agent(agent_type: str, input_data: Dict, tenant: str) -> Dict:
    """Invoke an agent."""
    return {
        "trace_id": "trace-001",
        "success": True,
        "output": {"message": "Agent executed successfully"}
    }


async def search_memory(query: str, memory_type: str, tenant: str, limit: int) -> list:
    """Search memory substrate."""
    return []


async def get_clear_metrics(environment: str) -> Dict[str, Any]:
    """Get CLEAR metrics."""
    return {
        "cost": {"total_tokens": 1500000, "estimated_cost_usd": 30.50},
        "latency": {"p50_ms": 150, "p95_ms": 450, "p99_ms": 1200},
        "efficacy": {"success_rate": 0.95, "avg_confidence": 0.87},
        "assurance": {"policy_compliance_rate": 0.99, "audit_coverage": 1.0},
        "reliability": {"uptime": 0.999, "error_rate": 0.01}
    }


if __name__ == '__main__':
    cli()
