services/api_gateway/main.py
"""
ANTS API Gateway.
Main entry point for all agent interactions.
"""
from contextlib import asynccontextmanager
from fastapi import FastAPI, HTTPException, Depends, Request
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field
from typing import Dict, Any, List, Optional
import structlog
import uuid

from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor


logger = structlog.get_logger()


# Request/Response Models
class AgentRequest(BaseModel):
    """Request to invoke an agent."""
    agent_type: str = Field(..., description="Type of agent to invoke")
    input_data: Dict[str, Any] = Field(..., description="Input data for agent")
    tenant_id: str = Field(..., description="Tenant identifier")
    user_id: Optional[str] = Field(None, description="User identifier")
    session_id: Optional[str] = Field(None, description="Session identifier")
    metadata: Dict[str, Any] = Field(default_factory=dict)


class AgentResponse(BaseModel):
    """Response from agent invocation."""
    trace_id: str
    success: bool
    output: Any
    actions_taken: List[Dict[str, Any]] = []
    latency_ms: float
    tokens_used: int = 0
    error: Optional[str] = None


class HealthResponse(BaseModel):
    """Health check response."""
    status: str
    version: str
    components: Dict[str, str]


# Application lifecycle
@asynccontextmanager
async def lifespan(app: FastAPI):
    """Application startup and shutdown."""
    # Startup
    logger.info("api_gateway_starting")

    # Initialize dependencies
    # - Database connections
    # - Agent registry
    # - Memory substrate
    # - Policy engine

    yield

    # Shutdown
    logger.info("api_gateway_stopping")


# Create FastAPI app
app = FastAPI(
    title="ANTS API Gateway",
    description="AI-Agent Native Tactical System API",
    version="1.0.0",
    lifespan=lifespan
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# OpenTelemetry instrumentation
FastAPIInstrumentor.instrument_app(app)


@app.get("/health", response_model=HealthResponse)
async def health_check():
    """Health check endpoint."""
    return HealthResponse(
        status="healthy",
        version="1.0.0",
        components={
            "database": "healthy",
            "memory": "healthy",
            "policy": "healthy",
            "agents": "healthy"
        }
    )


@app.post("/v1/agents/invoke", response_model=AgentResponse)
async def invoke_agent(request: AgentRequest):
    """
    Invoke an agent with the given input.
    """
    trace_id = str(uuid.uuid4())

    logger.info(
        "agent_invocation_started",
        trace_id=trace_id,
        agent_type=request.agent_type,
        tenant_id=request.tenant_id
    )

    try:
        # Get agent from registry
        # agent = agent_registry.get(request.agent_type)

        # Create context
        # context = AgentContext(
        #     trace_id=trace_id,
        #     tenant_id=request.tenant_id,
        #     user_id=request.user_id,
        #     session_id=request.session_id,
        #     metadata=request.metadata
        # )

        # Run agent
        # result = await agent.run(request.input_data, context)

        # Placeholder response
        return AgentResponse(
            trace_id=trace_id,
            success=True,
            output={"message": "Agent executed successfully"},
            actions_taken=[],
            latency_ms=100.0,
            tokens_used=500
        )

    except Exception as e:
        logger.error(
            "agent_invocation_failed",
            trace_id=trace_id,
            error=str(e)
        )

        raise HTTPException(
            status_code=500,
            detail=f"Agent invocation failed: {str(e)}"
        )


@app.get("/v1/agents", response_model=List[Dict[str, Any]])
async def list_agents():
    """List available agents."""
    return [
        {
            "type": "finance.reconciliation",
            "name": "Reconciliation Agent",
            "description": "Automates financial reconciliation"
        },
        {
            "type": "retail.inventory",
            "name": "Inventory Agent",
            "description": "Manages inventory levels and replenishment"
        },
        {
            "type": "cybersecurity.defender",
            "name": "Defender Triage Agent",
            "description": "Triages security alerts"
        }
    ]


@app.get("/v1/memory/{tenant_id}/search")
async def search_memory(
    tenant_id: str,
    query: str,
    memory_type: str = "semantic",
    limit: int = 10
):
    """Search agent memory."""
    # Placeholder
    return {
        "results": [],
        "total": 0,
        "query": query,
        "memory_type": memory_type
    }


@app.get("/v1/metrics/clear")
async def get_clear_metrics():
    """Get CLEAR metrics for all agents."""
    return {
        "cost": {
            "total_tokens": 1500000,
            "estimated_cost_usd": 30.50
        },
        "latency": {
            "p50_ms": 150,
            "p95_ms": 450,
            "p99_ms": 1200
        },
        "efficacy": {
            "success_rate": 0.95,
            "avg_confidence": 0.87
        },
        "assurance": {
            "policy_compliance_rate": 0.99,
            "audit_coverage": 1.0
        },
        "reliability": {
            "uptime": 0.999,
            "error_rate": 0.01
        }
    }


def run():
    """Run the API gateway."""
    import uvicorn
    uvicorn.run(
        "services.api_gateway.main:app",
        host="0.0.0.0",
        port=8000,
        reload=True
    )


if __name__ == "__main__":
    run()
