pyproject.toml
[project]
name = "ants"
version = "1.0.0"
description = "AI-Agent Native Tactical System for Enterprise"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Dwiref Sharma"}
]
keywords = ["ai", "agents", "enterprise", "azure", "nvidia"]

dependencies = [
    # Core Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",

    # Azure SDK
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-eventhub>=5.11.0",
    "azure-storage-blob>=12.19.0",
    "azure-search-documents>=11.4.0",
    "azure-cosmos>=4.5.0",
    "azure-digitaltwins-core>=1.2.0",
    "azure-ai-ml>=1.12.0",
    "azure-monitor-opentelemetry>=1.2.0",

    # LangChain Ecosystem
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.20",

    # Agent Frameworks
    "autogen>=0.2.0",
    "crewai>=0.1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.4",
    "redis>=5.0.0",

    # Vector Stores
    "weaviate-client>=4.4.0",
    "pymilvus>=2.3.0",

    # ML/AI
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",

    # NVIDIA
    "tritonclient[all]>=2.41.0",
    "nemo-guardrails>=0.7.0",

    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",

    # Policy
    "opa-python-client>=1.3.0",

    # Utilities
    "tenacity>=8.2.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "httpx>=0.26.0",
    "respx>=0.20.0",
]

rapids = [
    "cudf-cu12>=24.02.0",
    "cuml-cu12>=24.02.0",
    "cugraph-cu12>=24.02.0",
]

databricks = [
    "databricks-sdk>=0.17.0",
    "delta-spark>=3.0.0",
    "mlflow>=2.10.0",
]

[project.scripts]
antsctl = "platform.bootstrap.antsctl:cli"
ants-server = "services.api_gateway.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "platform", "services"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src", "platform", "services"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]
